refactor: move getFilePreviewConfig to file_explorer/utils, delete FilePreviewSheet
