--- front/components/spaces/FilePreviewSheet.tsx
@@ -1,501 +1,0 @@
-import { FrameRenderer } from "@app/components/assistant/conversation/interactive_content/FrameRenderer";
-import { clientFetch } from "@app/lib/egress/client";
-import type { ProcessedContent } from "@app/lib/file_content_utils";
-import { processFileContent } from "@app/lib/file_content_utils";
-import { getFileTypeIcon } from "@app/lib/file_icon_utils";
-import {
-  getFileDownloadUrl,
-  getFileProcessedUrl,
-  getFileViewUrl,
-  useFileContent,
-  useFileSignedUrl,
-} from "@app/lib/swr/files";
-import {
-  getFileFormatCategory,
-  isInteractiveContentType,
-  isMarkdownContentType,
-  isPdfContentType,
-} from "@app/types/files";
-
-/** Minimal file shape for preview (e.g. conversation files). Only sId, fileName, contentType are used by the renderer. */
-export type MinimalFileForPreview = {
-  sId: string;
-  fileName: string;
-  contentType: string;
-  /** When set (e.g. project knowledge tab), interactive frames resolve in this project space. */
-  projectId?: string | null;
-};
-
-import { assertNeverAndIgnore } from "@app/types/shared/utils/assert_never";
-import type { WorkspaceType } from "@app/types/user";
-import {
-  Button,
-  Clipboard,
-  ClipboardCheck,
-  CodeBlock,
-  Download01,
-  LinkExternal01,
-  Markdown,
-  Sheet,
-  SheetContainer,
-  SheetContent,
-  SheetHeader,
-  SheetTitle,
-  Spinner,
-  useCopyToClipboard,
-} from "@dust-tt/sparkle";
-// biome-ignore lint/correctness/noUnusedImports: ignored using `--suppress`
-import React, { useEffect, useState } from "react";
-
-/**
- * Content types compatible with the external viewer (currently Microsoft Office Online).
- * These files can be previewed using the viewer iframe.
- */
-const VIEWER_CONTENT_TYPES = [
-  "application/msword",
-  "application/vnd.openxmlformats-officedocument.wordprocessingml.document",
-  "application/vnd.ms-powerpoint",
-  "application/vnd.openxmlformats-officedocument.presentationml.presentation",
-  "application/vnd.ms-excel",
-  "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet",
-] as const;
-
-function isViewerCompatible(contentType: string): boolean {
-  return VIEWER_CONTENT_TYPES.includes(
-    contentType as (typeof VIEWER_CONTENT_TYPES)[number]
-  );
-}
-
-type FilePreviewCategory =
-  | "frame"
-  | "code"
-  | "pdf"
-  | "viewer"
-  | "audio"
-  | "markdown"
-  | "delimited"
-  | "text"
-  | "image";
-
-interface FilePreviewConfig {
-  category: FilePreviewCategory;
-  needsProcessedVersion: boolean;
-  supportsExternalViewer: boolean;
-  supportsCopyContent: boolean;
-}
-
-export function getFilePreviewConfig(contentType: string): FilePreviewConfig {
-  const category = getFileFormatCategory(contentType);
-
-  if (isInteractiveContentType(contentType)) {
-    return {
-      category: "frame",
-      needsProcessedVersion: false,
-      supportsExternalViewer: false,
-      supportsCopyContent: false,
-    };
-  }
-
-  if (isPdfContentType(contentType)) {
-    return {
-      category: "pdf",
-      needsProcessedVersion: true,
-      supportsExternalViewer: true,
-      supportsCopyContent: false,
-    };
-  }
-
-  if (isViewerCompatible(contentType)) {
-    return {
-      category: "viewer",
-      needsProcessedVersion: true,
-      supportsExternalViewer: true,
-      supportsCopyContent: false,
-    };
-  }
-
-  if (isMarkdownContentType(contentType)) {
-    return {
-      category: "markdown",
-      needsProcessedVersion: false,
-      supportsExternalViewer: false,
-      supportsCopyContent: true,
-    };
-  }
-
-  if (category === "code" || category === "data") {
-    return {
-      category: "code",
-      needsProcessedVersion: false,
-      supportsExternalViewer: false,
-      supportsCopyContent: true,
-    };
-  }
-
-  if (category === "audio") {
-    return {
-      category: "audio",
-      needsProcessedVersion: true,
-      supportsExternalViewer: false,
-      supportsCopyContent: false,
-    };
-  }
-
-  if (category === "delimited") {
-    return {
-      category: "delimited",
-      needsProcessedVersion: false,
-      supportsExternalViewer: false,
-      supportsCopyContent: false,
-    };
-  }
-
-  if (category === "image") {
-    return {
-      category: "image",
-      needsProcessedVersion: false,
-      supportsExternalViewer: false,
-      supportsCopyContent: false,
-    };
-  }
-
-  return {
-    category: "text",
-    needsProcessedVersion: false,
-    supportsExternalViewer: false,
-    supportsCopyContent: true,
-  };
-}
-
-function TextContent({ text }: { text: string }) {
-  return <Markdown content={text} isStreaming={false} />;
-}
-
-function AudioFileRenderer({
-  content,
-  audioUrl,
-}: {
-  content: ProcessedContent;
-  audioUrl: string;
-}) {
-  return (
-    <div className="flex flex-col gap-4">
-      <audio controls className="w-full" src={audioUrl}>
-        Your browser does not support the audio element.
-      </audio>
-      {content.text && (
-        <div className="flex flex-col gap-2">
-          <h4 className="text-sm font-semibold text-muted-foreground dark:text-muted-foreground-night">
-            Transcript
-          </h4>
-          <TextContent text={content.text} />
-        </div>
-      )}
-      {!content.text && (
-        <p className="text-sm text-muted-foreground dark:text-muted-foreground-night">
-          No transcript available.
-        </p>
-      )}
-    </div>
-  );
-}
-
-interface FileContentRendererProps {
-  file: MinimalFileForPreview;
-  owner: WorkspaceType;
-  previewConfig: FilePreviewConfig;
-  rawFileContent: string | null;
-  processedContent: ProcessedContent | null;
-  viewerSignedUrl: string | null;
-  viewerSignedUrlError: boolean;
-}
-
-function getViewerUrl(signedUrl: string): string {
-  return `https://view.officeapps.live.com/op/embed.aspx?src=${encodeURIComponent(signedUrl)}`;
-}
-
-function FileContentRenderer({
-  file,
-  owner,
-  previewConfig,
-  rawFileContent,
-  processedContent,
-  viewerSignedUrl,
-  viewerSignedUrlError,
-}: FileContentRendererProps) {
-  switch (previewConfig.category) {
-    case "pdf":
-      return (
-        <iframe
-          src={`${getFileViewUrl(owner, file.sId)}#navpanes=0`}
-          className="h-full min-h-[600px] w-full rounded-lg border-0"
-          title={file.fileName}
-        />
-      );
-
-    case "viewer":
-      if (viewerSignedUrlError && rawFileContent) {
-        return <TextContent text={rawFileContent} />;
-      }
-      if (viewerSignedUrl) {
-        return (
-          <iframe
-            src={getViewerUrl(viewerSignedUrl)}
-            className="h-full min-h-[600px] w-full rounded-lg border-0"
-            title={file.fileName}
-          />
-        );
-      }
-      return null;
-
-    case "audio":
-      return (
-        <AudioFileRenderer
-          content={processedContent ?? { text: "", format: "audio" }}
-          audioUrl={getFileViewUrl(owner, file.sId)}
-        />
-      );
-    case "image":
-      return (
-        <img
-          src={getFileViewUrl(owner, file.sId)}
-          alt={file.fileName}
-          className="max-h-[80vh] w-full rounded-lg object-contain"
-        />
-      );
-
-    case "frame":
-      return (
-        <FrameRenderer
-          fileId={file.sId}
-          projectId={file.projectId ?? null}
-          owner={owner}
-          lastEditedByAgentConfigurationId={undefined}
-          contentHash={undefined}
-        />
-      );
-    case "code":
-      return (
-        <CodeBlock
-          className="language-json max-h-60 overflow-y-auto"
-          wrapLongLines={true}
-        >
-          {rawFileContent}
-        </CodeBlock>
-      );
-    case "markdown":
-    case "delimited":
-    case "text":
-      if (processedContent) {
-        return <TextContent text={processedContent.text} />;
-      }
-      return null;
-    default:
-      assertNeverAndIgnore(previewConfig.category);
-  }
-}
-
-interface FilePreviewSheetProps {
-  owner: WorkspaceType;
-  file: MinimalFileForPreview | null;
-  isOpen: boolean;
-  onOpenChange: (open: boolean) => void;
-}
-
-// TODO(20260504 FILE SYSTEM) Replace this with FilePreviewDialog.
-export function FilePreviewSheet({
-  file,
-  owner,
-  isOpen,
-  onOpenChange,
-}: FilePreviewSheetProps) {
-  const previewConfig = getFilePreviewConfig(file?.contentType ?? "");
-  const isViewer = previewConfig.category === "viewer";
-
-  const {
-    signedUrl: viewerSignedUrl,
-    isLoading: isViewerSignedUrlLoading,
-    error: viewerSignedUrlError,
-  } = useFileSignedUrl({
-    fileId: file?.sId ?? null,
-    owner,
-    config: { disabled: !isOpen || !file || !isViewer },
-  });
-
-  const [isCopied, copy] = useCopyToClipboard();
-
-  const handleDownload = () => {
-    if (file) {
-      window.open(getFileDownloadUrl(owner, file.sId), "_blank");
-    }
-  };
-
-  const handleOpenInBrowser = () => {
-    if (file) {
-      if (isViewer && viewerSignedUrl) {
-        window.open(getViewerUrl(viewerSignedUrl), "_blank");
-      } else {
-        window.open(getFileViewUrl(owner, file.sId), "_blank");
-      }
-    }
-  };
-
-  // Fetch processed content directly (bypasses SWR to avoid caching Response
-  // objects whose body can only be read once).
-  const [processedText, setProcessedText] = useState<string | null>(null);
-  const [isProcessedTextLoaded, setIsProcessedTextLoaded] = useState(false);
-  useEffect(() => {
-    if (!isOpen || !file || !previewConfig.needsProcessedVersion) {
-      return;
-    }
-    const fileId = file.sId;
-    let cancelled = false;
-    setProcessedText(null);
-    setIsProcessedTextLoaded(false);
-
-    void (async () => {
-      const response = await clientFetch(getFileProcessedUrl(owner, fileId), {
-        redirect: "manual",
-      });
-      if (cancelled) {
-        return;
-      }
-      if (response.type !== "opaqueredirect" && response.ok) {
-        const text = await response.text();
-        if (!cancelled) {
-          setProcessedText(text);
-        }
-      }
-      if (!cancelled) {
-        setIsProcessedTextLoaded(true);
-      }
-    })();
-
-    return () => {
-      cancelled = true;
-    };
-  }, [isOpen, file, previewConfig.needsProcessedVersion, owner]);
-
-  const { fileContent: originalContent, error: originalContentError } =
-    useFileContent({
-      fileId: file?.sId ?? null,
-      owner,
-      config: {
-        disabled:
-          !isOpen ||
-          !file ||
-          previewConfig.needsProcessedVersion ||
-          previewConfig.category === "image",
-      },
-    });
-
-  const isPdf = previewConfig.category === "pdf";
-
-  const rawFileContent = previewConfig.needsProcessedVersion
-    ? processedText
-    : (originalContent ?? null);
-
-  const hasError =
-    !previewConfig.needsProcessedVersion && !!originalContentError;
-  const isImage = previewConfig.category === "image";
-  const isContentLoading =
-    isOpen &&
-    file &&
-    !hasError &&
-    !isPdf &&
-    !isViewer &&
-    !isImage &&
-    (previewConfig.needsProcessedVersion
-      ? !isProcessedTextLoaded
-      : !rawFileContent);
-  const isViewerLoading =
-    isOpen && file && isViewer && isViewerSignedUrlLoading;
-
-  const processedContent =
-    rawFileContent && file
-      ? processFileContent(rawFileContent, file.contentType)
-      : null;
-
-  const canCopyContent =
-    previewConfig.supportsCopyContent && !!processedContent?.text;
-
-  const renderContent = () => {
-    if (isContentLoading || isViewerLoading) {
-      return <Spinner />;
-    }
-
-    if (hasError) {
-      return (
-        <div className="flex h-48 w-full items-center justify-center text-muted-foreground dark:text-muted-foreground-night">
-          <p>Unable to preview this file. You can download it instead.</p>
-        </div>
-      );
-    }
-
-    if (!file) {
-      return null;
-    }
-
-    return (
-      <FileContentRenderer
-        file={file}
-        owner={owner}
-        previewConfig={previewConfig}
-        rawFileContent={rawFileContent}
-        processedContent={processedContent}
-        viewerSignedUrl={viewerSignedUrl}
-        viewerSignedUrlError={!!viewerSignedUrlError}
-      />
-    );
-  };
-
-  return (
-    <Sheet open={isOpen} onOpenChange={onOpenChange}>
-      <SheetContent size="xl">
-        <SheetHeader>
-          <SheetTitle
-            icon={
-              file
-                ? getFileTypeIcon(file.contentType, file.fileName)
-                : undefined
-            }
-          >
-            <div className="flex w-full items-center gap-2">
-              <span className="flex-1 truncate">{file?.fileName}</span>
-              {file && (
-                <div className="flex items-center gap-2">
-                  {canCopyContent && (
-                    <Button
-                      tooltip={isCopied ? "Copied!" : "Copy to clipboard"}
-                      variant="outline"
-                      size="icon-xs"
-                      icon={isCopied ? ClipboardCheck : Clipboard}
-                      onClick={() => copy(processedContent?.text ?? "")}
-                    />
-                  )}
-                  <Button
-                    variant="outline"
-                    size="icon-xs"
-                    icon={Download01}
-                    tooltip="Download"
-                    onClick={handleDownload}
-                  />
-                  {previewConfig.supportsExternalViewer && (
-                    <Button
-                      variant="outline"
-                      size="icon-xs"
-                      icon={LinkExternal01}
-                      tooltip="Open in browser"
-                      onClick={handleOpenInBrowser}
-                    />
-                  )}
-                </div>
-              )}
-            </div>
-          </SheetTitle>
-        </SheetHeader>
-        <SheetContainer>{renderContent()}</SheetContainer>
-      </SheetContent>
-    </Sheet>
-  );
-}
